Move code from .h to .cpp file
